util: Move CloudContext to util
