+++ util/storage.go
@@ -1,13 +1,6 @@
 package util
 
-import (
-	"net/http"
-
-	"golang.org/x/net/context"
-	"golang.org/x/oauth2/google"
-	"google.golang.org/appengine"
-	"google.golang.org/cloud"
-)
+import "google.golang.org/appengine"
 
 var FileNames = map[string]string{
 	"py":   "app.py",
@@ -16,25 +9,7 @@ var FileNames = map[string]string{
 	"java": "Application.java",
 }
 
-var cloudClient *http.Client
-
-const projID = "coduno"
 const TemplateBucket = "coduno-templates"
-
-func init() {
-	var err error
-	cloudClient, err = google.DefaultClient(context.Background())
-	if err != nil {
-		panic(err)
-	}
-}
-
-func CloudContext(parent context.Context) context.Context {
-	if parent == nil {
-		return cloud.NewContext(projID, cloudClient)
-	}
-	return cloud.WithContext(parent, projID, cloudClient)
-}
 
 func SubmissionBucket() string {
 	if appengine.IsDevAppServer() {
